docs(website): new website homepage
