+++ website/package.json
@@ -21,7 +21,8 @@
     "clsx": "^1.2.1",
     "prism-react-renderer": "^1.3.5",
     "react": "^17.0.2",
-    "react-dom": "^17.0.2"
+    "react-dom": "^17.0.2",
+    "three": "^0.144.0"
   },
   "devDependencies": {
     "@docusaurus/module-type-aliases": "^2.1.0",
